add a common .c file for all comrv demos
